fix: ProductRequest의 price 타입을 Long에서 Int로 변경

+++ app/src/main/java/woowacourse/shopping/data/model/request/ProductRequest.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ProductRequest(
     @SerialName("name") val name: String,
-    @SerialName("price") val price: Long,
+    @SerialName("price") val price: Int,
     @SerialName("imageUrl") val imageUrl: String,
     @SerialName("category") val category: String,
 )
